refactor: import modularity of graph operations
