+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-known_third_party = joblib,matplotlib,numpy,pandas,pysam,scipy,sourmash,tqdm
+known_third_party = joblib,matplotlib,numpy,pandas,pysam,sourmash,tqdm
