+++ amira/__main__.py
@@ -8,13 +8,16 @@ import time
 
 import matplotlib
 import matplotlib.pyplot as plt
-import numpy as np
-from joblib import Parallel, delayed
-from scipy.signal import find_peaks, savgol_filter
 from tqdm import tqdm
 
 from amira.__init__ import __version__
-from amira.construct_graph import GeneMerGraph, build_graph, merge_graphs
+from amira.construct_graph import GeneMerGraph
+from amira.graph_operations import (
+    build_multiprocessed_graph,
+    choose_kmer_size,
+    iterative_bubble_popping,
+    plot_node_coverages,
+)
 from amira.pre_process import convert_pandora_output, process_pandora_json
 
 matplotlib.use("Agg")
@@ -158,28 +161,6 @@ def get_options() -> argparse.Namespace:
     return args
 
 
-def build_multiprocessed_graph(annotatedReads, geneMer_size, cores, gene_positions=None):
-    batches = [set(list(annotatedReads.keys())[i::cores]) for i in range(cores)]
-    if gene_positions is not None:
-        sub_graphs = Parallel(n_jobs=cores)(
-            delayed(build_graph)(
-                {k: annotatedReads[k] for k in annotatedReads if k in batch},
-                geneMer_size,
-                {k: gene_positions[k] for k in gene_positions if k in batch},
-            )
-            for batch in batches
-        )
-    else:
-        sub_graphs = Parallel(n_jobs=cores)(
-            delayed(build_graph)(
-                {k: annotatedReads[k] for k in annotatedReads if k in batch}, geneMer_size
-            )
-            for batch in batches
-        )
-    merged_graph = merge_graphs(sub_graphs)
-    return merged_graph
-
-
 def plot_read_length_distribution(annotatedReads, output_dir):
     read_lengths = []
     for read in annotatedReads:
@@ -209,80 +190,6 @@ def write_debug_files(
         os.path.join(output_dir, "pre_correction_gene_mer_graph"), geneMer_size, 1, 1
     )
     return raw_graph
-
-
-def plot_node_coverages(unitig_coverages, filename):
-    # Calculate the frequency of each coverage value with bins of width 5
-    max_coverage = max(unitig_coverages)
-    # Create bins with a step size of 5
-    bins = np.arange(0, max_coverage + 5, 5)  # Bins of width 5
-    hist, bin_edges = np.histogram(unitig_coverages, bins=bins)
-    # Midpoints of bins for plotting
-    x_values = 0.5 * (bin_edges[:-1] + bin_edges[1:])
-    # Apply log transformation to the counts, adding 1 to avoid log(0)
-    log_counts = np.log(hist + 1)
-    # Smooth the log-transformed histogram counts using a Savitzky-Golay filter
-    window_length, poly_order = 31, 5  # Window length must be odd
-    if len(log_counts) < window_length:
-        window_length = max(5, len(log_counts) // 2 * 2 - 1)  # Smallest odd number >= 3
-    smoothed_log_counts = savgol_filter(log_counts, window_length, poly_order)
-
-    # Plot histogram
-    plt.figure(figsize=(10, 6))
-    plt.bar(
-        x_values,
-        log_counts,
-        width=5,
-        label="Counts",
-        color="white",
-        edgecolor="black",
-        align="center",
-    )
-    plt.plot(x_values, smoothed_log_counts, color="red", label="Smoothed counts")
-    plt.title("Histogram of mean unitig coverages with Smoothed Curve")
-    plt.xlabel("Unitig Coverage")
-    plt.ylabel("Log of absolute frequency")
-    plt.xlim([0, max(x_values) + 5])
-    plt.legend()
-    plt.savefig(filename)
-    plt.close()
-
-    # Identify peaks and troughs
-    peaks, _ = find_peaks(
-        [min(smoothed_log_counts)] + list(smoothed_log_counts),
-        [min(smoothed_log_counts)],
-        prominence=0.1,
-    )
-    peaks = peaks - 1
-    first_peak_index = np.where(x_values == x_values[peaks[0]])[0][0]
-    second_peak_index = np.where(x_values == x_values[peaks[1]])[0][0]
-    trough_index = (
-        np.argmin(smoothed_log_counts[first_peak_index : second_peak_index + 1]) + first_peak_index
-    )
-    trough_value = x_values[trough_index]
-
-    # Plot the histogram and the trough
-    plt.figure(figsize=(10, 6))
-    plt.bar(
-        x_values,
-        log_counts,
-        width=5,
-        label="Counts",
-        color="white",
-        edgecolor="black",
-        align="center",
-    )
-    plt.plot(x_values, smoothed_log_counts, color="red", label="Smoothed counts")
-    plt.axvline(x=trough_value, color="r", linestyle="--", label=f"Trough at x={trough_value:.2f}")
-    plt.title("Histogram of node coverages with Smoothed Curve")
-    plt.xlabel("Node Coverage")
-    plt.ylabel("Log of absolute frequency")
-    plt.xlim([0, max(x_values) + 5])
-    plt.legend()
-    plt.savefig(filename)
-    plt.close()
-
-    return trough_value
 
 
 def estimate_copy_number(amira_allele, copy_numbers_per_component, additional_copy_numbers):
@@ -381,7 +288,7 @@ def write_fastq(fastq_file, data):
             fh.write(f'{value["quality"]}\n')
 
 
-def process_reference_alelles(path_to_interesting_genes):
+def process_reference_alleles(path_to_interesting_genes):
     # import the list of genes of interest
     with open(path_to_interesting_genes, "r") as i:
         reference_content = i.read().split(">")[1:]
@@ -399,112 +306,6 @@ def process_reference_alelles(path_to_interesting_genes):
             reference_alleles[gene_name] = {}
         reference_alleles[gene_name][allele_name] = sequence
     return reference_alleles, genesOfInterest
-
-
-def choose_kmer_size(
-    overall_mean_node_coverage,
-    new_annotatedReads,
-    cores,
-    new_gene_position_dict,
-    sample_genesOfInterest,
-):
-    geneMer_size = 3
-    if overall_mean_node_coverage >= 20:
-        for k in range(3, 16, 2):
-            # Build the graph with the current k value
-            graph = build_multiprocessed_graph(
-                new_annotatedReads.copy(), k, cores, new_gene_position_dict.copy()
-            )
-
-            def is_component_valid(component):
-                amr_nodes = {
-                    n.__hash__()
-                    for g in sample_genesOfInterest
-                    for n in graph.get_nodes_containing(g)
-                }
-                nodes_in_component = [n.__hash__() for n in graph.get_nodes_in_component(component)]
-                reads = graph.collect_reads_in_path(
-                    [n for n in nodes_in_component if n in amr_nodes]
-                )
-                lengths = [len(graph.get_reads()[r]) for r in reads]
-                if len(lengths) != 0:
-                    return (
-                        len([length for length in lengths if length >= (2 * k - 1)]) / len(lengths)
-                        >= 0.8
-                    )
-                else:
-                    return True
-
-            if all(is_component_valid(c) for c in graph.components()):
-                geneMer_size = k
-            else:
-                break
-    return geneMer_size
-
-
-def iterative_bubble_popping(
-    new_annotatedReads,
-    new_gene_position_dict,
-    cleaning_iterations,
-    geneMer_size,
-    cores,
-    short_reads,
-    short_read_gene_positions,
-    fastq_content,
-    output_dir,
-    node_min_coverage,
-    sample_genesOfInterest,
-    min_path_coverage,
-):
-    prev_nodes = 0
-    components_to_skip = set()
-    for this_iteration in range(cleaning_iterations):
-        sys.stderr.write(
-            f"\nAmira: running graph cleaning iteration {this_iteration+1}/{cleaning_iterations}\n"
-        )
-        graph = build_multiprocessed_graph(
-            new_annotatedReads, geneMer_size, cores, new_gene_position_dict
-        )
-        graph.filter_graph(node_min_coverage, 1)
-        new_annotatedReads, new_gene_position_dict = graph.correct_reads(fastq_content)
-        graph = build_multiprocessed_graph(
-            new_annotatedReads, geneMer_size, cores, new_gene_position_dict
-        )
-        # check if the current number of nodes is equal to the previous number of nodes
-        if len(graph.get_nodes()) == prev_nodes:
-            sys.stderr.write(f"\n\tAmira: terminating cleaning at iteration {this_iteration+1}\n")
-            break
-        prev_nodes = len(graph.get_nodes())
-        sys.stderr.write("\n\tAmira: removing dead ends\n")
-        # collect the reads that have fewer than k genes
-        short_reads.update(graph.get_short_read_annotations())
-        short_read_gene_positions.update(graph.get_short_read_gene_positions())
-        graph.remove_short_linear_paths(geneMer_size)
-        new_annotatedReads, new_gene_position_dict = graph.correct_reads(fastq_content)
-        sys.stderr.write("\n\tAmira: popping bubbles using 1 CPU\n")
-        graph = build_multiprocessed_graph(
-            new_annotatedReads, geneMer_size, cores, new_gene_position_dict
-        )
-        # collect the reads that have fewer than k genes
-        short_reads.update(graph.get_short_read_annotations())
-        short_read_gene_positions.update(graph.get_short_read_gene_positions())
-        graph.generate_gml(
-            os.path.join(output_dir, "intermediate_graph"),
-            geneMer_size,
-            node_min_coverage,
-            1,
-        )
-        new_annotatedReads, new_gene_position_dict, path_coverages, min_path_coverage = (
-            graph.correct_low_coverage_paths(
-                fastq_content,
-                sample_genesOfInterest,
-                cores,
-                min_path_coverage,
-                components_to_skip,
-                True,
-            )
-        )
-    return new_annotatedReads, new_gene_position_dict
 
 
 def get_found_genes(clusters_of_interest):
@@ -595,7 +396,7 @@ def main() -> None:
     if not os.path.exists(args.output_dir):
         os.mkdir(args.output_dir)
     # import the list of genes of interest
-    reference_alleles, genesOfInterest = process_reference_alelles(args.path_to_interesting_genes)
+    reference_alleles, genesOfInterest = process_reference_alleles(args.path_to_interesting_genes)
     # import a JSON of genes on reads
     if args.pandoraJSON:
         if not args.quiet:
@@ -613,7 +414,7 @@ def main() -> None:
         pandora_consensus = parse_fastq(args.pandoraConsensus)
         # randomly sample 100,000 reads
         if args.sample_reads:
-            annotatedReads = downsample_reads(annotatedReads, 100000)
+            annotatedReads = downsample_reads(annotatedReads, 10000)
     # load the pandora consensus and convert to a dictionary
     if args.pandoraSam:
         if not args.quiet:
