+++ amira/construct_graph.py
@@ -23,99 +23,6 @@ from amira.construct_read import Read
 sys.setrecursionlimit(50000)
 
 
-def build_graph(read_dict, kmer_size, gene_positions=None):
-    graph = GeneMerGraph(read_dict, kmer_size, gene_positions)
-    return graph
-
-
-def merge_nodes(sub_graphs, fastq_data=None):
-    reference_graph = sub_graphs[0]
-    # iterate through the subgraphs
-    for graph in sub_graphs[1:]:
-        # iterate through the reads in the subgraph
-        for read_id in graph.get_readNodes():
-            # get the nodes on each read
-            node_hashes_on_read = graph.get_readNodes()[read_id]
-            # get the direction of each node on each read
-            node_directions_on_read = graph.get_readNodeDirections()[read_id]
-            # get the positions of each node on each read
-            node_positions_on_read = graph.get_readNodePositions()[read_id]
-            # iterate through the nodes
-            for i in range(len(node_hashes_on_read)):
-                # get the node object
-                node_in_subgraph = graph.get_node_by_hash(node_hashes_on_read[i])
-                # add the node to the reference graph
-                node_in_reference = reference_graph.add_node(
-                    node_in_subgraph.get_geneMer(), node_in_subgraph.get_reads()
-                )
-                # add to the minhash
-                if fastq_data is not None:
-                    if node_in_reference.get_minhash() is None:
-                        node_in_reference.set_minhash(node_in_subgraph.get_minhash())
-                    node_in_reference.get_minhash().add_many(node_in_subgraph.get_minhash())
-                # increment the node coverage
-                node_in_reference.increment_node_coverage()
-                # add the node to the read in the reference graph
-                reference_graph.add_node_to_read(
-                    node_in_reference,
-                    read_id,
-                    node_directions_on_read[i],
-                    node_positions_on_read[i],
-                )
-    return reference_graph
-
-
-def merge_edges(sub_graphs, reference_graph):
-    for graph in sub_graphs[1:]:
-        for edge_hash in graph.get_edges():
-            if edge_hash not in reference_graph.get_edges():
-                # get the edge object in the subgraph
-                edge_in_subgraph = graph.get_edge_by_hash(edge_hash)
-                # change the source and target node to those in the reference graph
-                reference_source_node = edge_in_subgraph.set_sourceNode(
-                    reference_graph.get_node_by_hash(edge_in_subgraph.get_sourceNode().__hash__())
-                )
-                edge_in_subgraph.set_targetNode(
-                    reference_graph.get_node_by_hash(edge_in_subgraph.get_targetNode().__hash__())
-                )
-                # add the edge to the node
-                reference_graph.add_edge_to_node(reference_source_node, edge_in_subgraph)
-                # add the edge object to the graph
-                reference_graph.get_edges()[edge_hash] = edge_in_subgraph
-            else:
-                # get the reference edge
-                reference_edge = reference_graph.get_edge_by_hash(edge_hash)
-                # extend the coverage
-                reference_edge.extend_edge_coverage(reference_edge.get_edge_coverage())
-
-
-def merge_reads(sub_graphs, reference_graph):
-    for graph in sub_graphs[1:]:
-        for read in graph.get_reads():
-            reference_graph.get_reads()[read] = graph.get_reads()[read]
-            if reference_graph.get_gene_positions() is not None:
-                reference_graph.get_gene_positions()[read] = graph.get_gene_positions()[read]
-        for read in graph.get_short_read_annotations():
-            reference_graph.get_short_read_annotations()[read] = graph.get_short_read_annotations()[
-                read
-            ]
-            if graph.get_gene_positions() is not None:
-                reference_graph.get_short_read_gene_positions()[
-                    read
-                ] = graph.get_short_read_gene_positions()[read]
-
-
-def merge_graphs(sub_graphs):
-    # merge the nodes
-    reference_graph = merge_nodes(sub_graphs)
-    # merge the edges
-    merge_edges(sub_graphs, reference_graph)
-    # merge the reads
-    merge_reads(sub_graphs, reference_graph)
-    reference_graph.assign_component_ids()
-    return reference_graph
-
-
 class GeneMerGraph:
     def __init__(self, readDict, kmerSize, gene_positions=None):
         self._reads = readDict
